requirements.txt
google-cloud-storage
tqdm
pandas
Pillow
azure
wget
boto3